Refactored video transcoding logic into custom hook
